third_party/blink/web_tests/external/wpt/editing/run/outdent_1-1000-expected.txt

This is a testharness.js-based test.
Found 52 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["outdent",""]] "<blockquote><p>foo[bar]</p><p>baz</p></blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foobar</p><blockquote><p>baz</p></blockquote><p>extra</p>" but got "foobar<br><blockquote><p>baz</p></blockquote><p>extra</p>"
[FAIL] [["outdent",""]] "<blockquote><p>foo[bar</p><p>b]az</p></blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foobar</p><p>baz</p><p>extra</p>" but got "foobar<br><p>baz</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["outdent",""]] "<blockquote style=\\"margin-right: 0px;\\" dir=\\"ltr\\"><p>foo[bar]</p><p>baz</p></blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foobar</p><blockquote><p>baz</p></blockquote><p>extra</p>" but got "foobar<br><blockquote style=\\"margin-right:0px\\" dir=\\"ltr\\"><p>baz</p></blockquote><p>extra</p>"
[FAIL] [["stylewithcss","true"],["outdent",""]] "<blockquote style=\\"margin-right: 0px;\\" dir=\\"ltr\\"><p>foo[bar]</p><p>baz</p></blockquote><p>extra" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["outdent",""]] "<blockquote style=\\"margin-right: 0px;\\" dir=\\"ltr\\"><p>foo[bar]</p><p>baz</p></blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foobar</p><blockquote><p>baz</p></blockquote><p>extra</p>" but got "foobar<br><blockquote style=\\"margin-right:0px\\" dir=\\"ltr\\"><p>baz</p></blockquote><p>extra</p>"
[FAIL] [["stylewithcss","true"],["outdent",""]] "<blockquote style=\\"margin-right: 0px;\\" dir=\\"ltr\\"><p>foo[bar</p><p>b]az</p></blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foobar</p><p>baz</p><p>extra</p>" but got "foobar<br><p>baz</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["outdent",""]] "<blockquote style=\\"margin-right: 0px;\\" dir=\\"ltr\\"><p>foo[bar</p><p>b]az</p></blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foobar</p><p>baz</p><p>extra</p>" but got "foobar<br><p>baz</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["outdent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px;\\"><p>foo[bar]</p><p>baz</p></blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"webkit-indent-blockquote\\"><p>foobar</p><blockquote><p>baz</p></blockquote></div><p>extra</p>" but got "foobar<br><blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>baz</p></blockquote><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["outdent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px;\\"><p>foo[bar]</p><p>baz</p></blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"webkit-indent-blockquote\\"><p>foobar</p><blockquote><p>baz</p></blockquote></div><p>extra</p>" but got "foobar<br><blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>baz</p></blockquote><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["outdent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px;\\"><p>foo[bar]</p><p>baz</p></blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"webkit-indent-blockquote\\"><p>foobar</p><blockquote><p>baz</p></blockquote></div><p>extra</p>" but got "foobar<br><blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>baz</p></blockquote><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["outdent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px;\\"><p>foo[bar]</p><p>baz</p></blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"webkit-indent-blockquote\\"><p>foobar</p><blockquote><p>baz</p></blockquote></div><p>extra</p>" but got "foobar<br><blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin:0 0 0 40px; border:none; padding:0px\\"><p>baz</p></blockquote><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["outdent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px;\\"><p>foo[bar</p><p>b]az</p></blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"webkit-indent-blockquote\\"><p>foobar</p><p>baz</p></div><p>extra</p>" but got "foobar<br><p>baz</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["outdent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px;\\"><p>foo[bar</p><p>b]az</p></blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"webkit-indent-blockquote\\"><p>foobar</p><p>baz</p></div><p>extra</p>" but got "foobar<br><p>baz</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["outdent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px;\\"><p>foo[bar</p><p>b]az</p></blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"webkit-indent-blockquote\\"><p>foobar</p><p>baz</p></div><p>extra</p>" but got "foobar<br><p>baz</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["outdent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px;\\"><p>foo[bar</p><p>b]az</p></blockquote><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"webkit-indent-blockquote\\"><p>foobar</p><p>baz</p></div><p>extra</p>" but got "foobar<br><p>baz</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["outdent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px;\\"><p>foo[bar]</p></blockquote><p>baz</p><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"webkit-indent-blockquote\\"><p>foobar</p></div><p>baz</p><p>extra</p>" but got "<p>foobar</p><p>baz</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["outdent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px;\\"><p>foo[bar]</p></blockquote><p>baz</p><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"webkit-indent-blockquote\\"><p>foobar</p></div><p>baz</p><p>extra</p>" but got "<p>foobar</p><p>baz</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["outdent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px;\\"><p>foo[bar]</p></blockquote><p>baz</p><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"webkit-indent-blockquote\\"><p>foobar</p></div><p>baz</p><p>extra</p>" but got "<p>foobar</p><p>baz</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["outdent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px;\\"><p>foo[bar]</p></blockquote><p>baz</p><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"webkit-indent-blockquote\\"><p>foobar</p></div><p>baz</p><p>extra</p>" but got "<p>foobar</p><p>baz</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","div"],["outdent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px;\\"><p>foo[bar</p></blockquote><p>b]az</p><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"webkit-indent-blockquote\\"><p>foobar</p></div><p>baz</p><p>extra</p>" but got "<p>foobar</p><p>baz</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["outdent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px;\\"><p>foo[bar</p></blockquote><p>b]az</p><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"webkit-indent-blockquote\\"><p>foobar</p></div><p>baz</p><p>extra</p>" but got "<p>foobar</p><p>baz</p><p>extra</p>"
[FAIL] [["stylewithcss","true"],["defaultparagraphseparator","p"],["outdent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px;\\"><p>foo[bar</p></blockquote><p>b]az</p><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"webkit-indent-blockquote\\"><p>foobar</p></div><p>baz</p><p>extra</p>" but got "<p>foobar</p><p>baz</p><p>extra</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["outdent",""]] "<blockquote class=\\"webkit-indent-blockquote\\" style=\\"margin: 0 0 0 40px; border: none; padding: 0px;\\"><p>foo[bar</p></blockquote><p>b]az</p><p>extra" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div class=\\"webkit-indent-blockquote\\"><p>foobar</p></div><p>baz</p><p>extra</p>" but got "<p>foobar</p><p>baz</p><p>extra</p>"
[FAIL] [["outdent",""]] "<blockquote><blockquote data-abc=def>foo[bar]baz</blockquote></blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote data-abc=\\"def\\">foobarbaz</blockquote>" but got "<blockquote>foobarbaz</blockquote>"
[FAIL] [["defaultparagraphseparator","div"],["outdent",""]] "<blockquote id=abc>foo[bar]baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div id=\\"abc\\">foobarbaz</div>" but got "foobarbaz"
[FAIL] [["defaultparagraphseparator","p"],["outdent",""]] "<blockquote id=abc>foo[bar]baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div id=\\"abc\\">foobarbaz</div>" but got "foobarbaz"
[FAIL] [["defaultparagraphseparator","div"],["outdent",""]] "<blockquote style=\\"color: blue\\">foo[bar]baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"color:rgb(0, 0, 255)\\">foobarbaz</div>" but got "foobarbaz"
[FAIL] [["defaultparagraphseparator","p"],["outdent",""]] "<blockquote style=\\"color: blue\\">foo[bar]baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"color:rgb(0, 0, 255)\\">foobarbaz</div>" but got "foobarbaz"
[FAIL] [["outdent",""]] "<blockquote><blockquote><p>foo[bar]<p>baz</blockquote></blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote><p>foobar</p><blockquote><p>baz</p></blockquote></blockquote>" but got "<blockquote>foobar<br><blockquote><p>baz</p></blockquote></blockquote>"
[FAIL] [["outdent",""]] "<blockquote><blockquote data-abc=def><p>foo[bar]<p>baz</blockquote></blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote data-abc=\\"def\\"><p>foobar</p><blockquote><p>baz</p></blockquote></blockquote>" but got "<blockquote>foobar<br><blockquote data-abc=\\"def\\"><p>baz</p></blockquote></blockquote>"
[FAIL] [["outdent",""]] "<blockquote data-abc=def><blockquote><p>foo[bar]<p>baz</blockquote></blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote data-abc=\\"def\\"><p>foobar</p><blockquote><p>baz</p></blockquote></blockquote>" but got "<blockquote data-abc=\\"def\\">foobar<br><blockquote><p>baz</p></blockquote></blockquote>"
[FAIL] [["outdent",""]] "<blockquote><div><p>foo[bar]<p>baz</div></blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><p>foobar</p><blockquote><p>baz</p></blockquote></div>" but got "foobar<br><blockquote><p>baz</p></blockquote>"
[FAIL] [["outdent",""]] "<blockquote><div id=abc><p>foo[bar]<p>baz</div></blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div id=\\"abc\\"><p>foobar</p><blockquote><p>baz</p></blockquote></div>" but got "foobar<br><blockquote><div id=\\"abc\\"><p>baz</p></div></blockquote>"
[FAIL] [["defaultparagraphseparator","div"],["outdent",""]] "<blockquote id=abc><p>foo[bar]<p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div id=\\"abc\\"><p>foobar</p><blockquote><p>baz</p></blockquote></div>" but got "foobar<br><blockquote id=\\"abc\\"><p>baz</p></blockquote>"
[FAIL] [["defaultparagraphseparator","p"],["outdent",""]] "<blockquote id=abc><p>foo[bar]<p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div id=\\"abc\\"><p>foobar</p><blockquote><p>baz</p></blockquote></div>" but got "foobar<br><blockquote id=\\"abc\\"><p>baz</p></blockquote>"
[FAIL] [["defaultparagraphseparator","div"],["outdent",""]] "<blockquote style=\\"color: blue\\"><p>foo[bar]<p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"color:rgb(0, 0, 255)\\"><p>foobar</p><blockquote><p>baz</p></blockquote></div>" but got "<span style=\\"color:rgb(0, 0, 255)\\">foobar</span><br><blockquote style=\\"color:rgb(0, 0, 255)\\"><p>baz</p></blockquote>"
[FAIL] [["defaultparagraphseparator","p"],["outdent",""]] "<blockquote style=\\"color: blue\\"><p>foo[bar]<p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"color:rgb(0, 0, 255)\\"><p>foobar</p><blockquote><p>baz</p></blockquote></div>" but got "<span style=\\"color:rgb(0, 0, 255)\\">foobar</span><br><blockquote style=\\"color:rgb(0, 0, 255)\\"><p>baz</p></blockquote>"
[FAIL] [["outdent",""]] "<blockquote><p><b>foo[bar]</b><p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><b>foobar</b></p><blockquote><p>baz</p></blockquote>" but got "<b>foobar</b><br><blockquote><p>baz</p></blockquote>"
[FAIL] [["outdent",""]] "<blockquote><p><strong>foo[bar]</strong><p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><strong>foobar</strong></p><blockquote><p>baz</p></blockquote>" but got "<strong>foobar</strong><br><blockquote><p>baz</p></blockquote>"
[FAIL] [["outdent",""]] "<blockquote><p><span>foo[bar]</span><p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span>foobar</span></p><blockquote><p>baz</p></blockquote>" but got "foobar<br><blockquote><p>baz</p></blockquote>"
[FAIL] [["outdent",""]] "<blockquote><blockquote style=\\"color: blue\\"><p>foo[bar]</blockquote><p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote style=\\"color:rgb(0, 0, 255)\\"><p>foobar</p></blockquote><blockquote><p>baz</p></blockquote>" but got "<blockquote><p>foobar</p></blockquote><blockquote><p>baz</p></blockquote>"
[FAIL] [["outdent",""]] "<blockquote style=\\"color: blue\\"><blockquote><p>foo[bar]</blockquote><p>baz</blockquote>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote style=\\"color:rgb(0, 0, 255)\\"><p>foobar</p><p>baz</p></blockquote>" but got "<blockquote style=\\"color:rgb(0, 0, 255)\\"><p>foobar</p></blockquote><blockquote style=\\"color:rgb(0, 0, 255)\\"><p>baz</p></blockquote>"
[FAIL] [["defaultparagraphseparator","div"],["outdent",""]] "<ol><li>foo<li>[bar]<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><div>bar</div><ol><li>baz</li></ol>" but got "<ol><li>foo</li></ol>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["outdent",""]] "<ol><li>foo<li>[bar]<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><p>bar</p><ol><li>baz</li></ol>" but got "<ol><li>foo</li></ol>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["outdent",""]] "<ol data-start=1 data-end=2><li>foo<li>bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><div>bar</div><ol><li>baz</li></ol>" but got "<ol><li>foo</li></ol>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["outdent",""]] "<ol data-start=1 data-end=2><li>foo<li>bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><p>bar</p><ol><li>baz</li></ol>" but got "<ol><li>foo</li></ol>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["outdent",""]] "<ol><li>[foo]<br>bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo<br>bar</div><ol><li>baz</li></ol>" but got "foo<br>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["outdent",""]] "<ol><li>[foo]<br>bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<br>bar</p><ol><li>baz</li></ol>" but got "foo<br>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["outdent",""]] "<ol><li>foo<br>[bar]<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo<br>bar</div><ol><li>baz</li></ol>" but got "foo<br>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["outdent",""]] "<ol><li>foo<br>[bar]<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<br>bar</p><ol><li>baz</li></ol>" but got "foo<br>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["outdent",""]] "<ol><li><div>[foo]</div>bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><div>bar</div><ol><li>baz</li></ol>" but got "<div style=\\"display:inline !important\\">foo</div>bar<br><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["outdent",""]] "<ol><li><div>[foo]</div>bar<li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><p>bar</p><ol><li>baz</li></ol>" but got "<div style=\\"display:inline !important\\">foo</div>bar<br><ol><li>baz</li></ol>"
Harness: the test ran to completion.
